renaming app component folder
