+++ frontend/src/App.tsx
@@ -7,10 +7,10 @@ import MatchingPage from "./FullPageRoutes/Matching"
 import { Box, makeStyles, Grid } from "@material-ui/core"
 import RecordPage from "./FullPageRoutes/RecordPage"
 import ProvideContext from "./components/utils/ProvideContext"
-import AppHeader from "./components/app/AppHeader"
-import AppFooter from "./components/app/AppFooter"
+import AppHeader from "./components/core/AppHeader"
+import AppFooter from "./components/core/AppFooter"
+import AppSideBar from "./components/core/AppSidebar"
 import { ProtectedRoute } from "./components/ProtectedRoute"
-import AppSideBar from "./components/app/AppSidebar"
 import MailPage from "./FullPageRoutes/MailPage"
 import VideoViewPage from "./FullPageRoutes/VideoViewPage"
 import Help from "./FullPageRoutes/Help"
