+++ frontend/src/components/core/AppFooter.tsx
